requirements.txt
google-generativeai==0.3.2
python-dotenv==1.0.0
requests==2.31.0
rich==13.7.0
flask==3.0.0
flask-cors==4.0.0
markdown2==2.4.12
gunicorn==21.2.0